consumer/requirements.txt
confluent-kafka==2.5.0
torch==2.3.1
transformers==4.44.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
